integrated sentry for logging errors
